Added persistence provider prepare check missing migrations
Ensures an error is logged when migrations are missing. This is done by delegating to `Sequel::Migrator` to check if migrations are current which means all local migration file names must match the file names listed in the `schema_migrations` table.

The `NoMethodError` check is for situations when a newly created database has *no migrations* and can be safely ignored otherwise you'd not be able to migrate at all. This happens because the `ROM::SQL.migration` method assumes you can always message `current_gateway` with `#migration` but will be `nil` in this situation.