Added RuboCop Metrics/BlockLength comments to persistence provider
Necessary to avoid RuboCop violations when generating a new Hanami project. The persistence code is definitely unruly and needs to be fixed but am waiting for the Hanami team to officially support the persistance layer. This is marked as a TODO for now and will be removed in the future.