Fixed core builder application configuration for CSP and assets
Discovered the Content Security Policy (CSP) for web manifests is not set by default so this fixes the CSP so it can load a local manifest to enable basic Progressive Web Application (PWA) support.

Due to recently adding a manifest for PWAs and an SVG icon, the Rack Static middleware was updated to serve these assets properly.

Milestone: patch