Added icons to main slice application layout
Necessary to provide basic Progressive Web Application (PWA) support and support for application icons and branding in general. Images are pulled from Alchemists as placeholders and meant to be customized. This also avoids having to add binary information to the Git repository which is a bad practice in general.

Language and viewport metadata was updated to be standards compliant.

Milestone: minor