Updated stylesheet to use a radial background and centered layout
Necessary to improve the default visual appearance of the home page, provide a better name for the stylesheet, and allow the stylesheet to be conveniently deleted if not desired. This will soon be used by the home slice show template.

The spec was updated to only check for file existence to reduce maintenance burden since no additional manipulation to the original template is performed.

Milestone: minor