Refactored main slice as home slice
Necessary to give this slice a more descriptive and specific name instead of `main` which might inadvertently encourage a dumping ground for mixed functionality when each slice should have a specific purpose. This also means that if this slice is not desired then removing it takes minimal effort.

Milestone: minor