Added title, icon, label, and info to home show template
Necessary to leverage the `content_for` helper so page titles can be dynamic. The new stylesheet also improves the appearance. Additional resources are provided in the footer so people can learn more as desired.

HTML comments are replaced with ERB values in order to avoid ERB evaluation by the builder since Hanami View can render properly once the skeleton is generated. A spec fixture is used to confirm this works properly.

The spec was updated to only check for the label since that is the only dynamic information injected into the template.

Milestone: minor