Fixed missing Puma temporary directory
When using this gem to create a new Hanami project, the `tmp` directory is created for you. There are situations where this doesn't work like when cloning the repository and launching the server or manually deleting the `tmp` directory and forgetting to recreate the directory. This change ensures that the temporary directory is created if it doesn't exist for these situations.

Milestone: patch