Updated Puma configuration builder to use pre-forking and new port
Necessary to match the defaults provided by Hanami but also ensure pre-forking only happens for positive concurrency. Otherwise, you'll experience the following errors:

    [__NSCFConstantString initialize] may have been in progress in another thread when fork() was called. We cannot safely call it or ignore it in the fork() child process. Crashing instead. Set a breakpoint on objc_initializeAfterForkError to debug.

The port was updated from 9050 to 2443 to be easier to remember since 443 is used for HTTPS and be closer aligned with the default, insecure, 2300 port.

Milestone: minor