Added asset extension
Necessary to ensure assets are compiled for newly built project skeletons. This helps reduce an additional manual step so people can get up and running quickly with a new application.

Node is being used instead of directly using `hanami assets compile` since we don't want to prematurely load the Hanami application and force it to validate any of the application settings (like database URL, for example) which will fail the building of the project skeleton.

Milestone: minor