Added layout builder flash alerts and notices
Necessary to provide default flash alert and notice messages. The tag helpers were also updated to use consistent syntax.

RuboCop was updated to ignore the metric issues because we need to keep all of this logic together. Breaking this up further would not be helpful in this case.

Milestone: minor